ai/image_generation/gemini_service.py
"""
Google Gemini image generation service implementation.
Handles image generation using Gemini AI models.
"""
import google.generativeai as genai
from typing import Dict, Any, Optional
import asyncio
import base64
import io
from PIL import Image
from loguru import logger

from .base import ImageGenerationBase
from config.settings import get_settings
from utils.constants import ERROR_MESSAGES, SUCCESS_MESSAGES


class GeminiImageService(ImageGenerationBase):
    """Google Gemini image generation service."""

    def __init__(self, **kwargs):
        """Initialize Gemini image generation service."""
        settings = get_settings()
        super().__init__(
            model_name="gemini",
            api_key=settings.gemini_api_key,
            model=settings.gemini_model,
            **kwargs
        )
        self.client = None
        self.model_instance = None

    async def initialize(self) -> None:
        """Initialize Gemini client and model."""
        try:
            if not self.config.get('api_key'):
                raise ValueError("Gemini API key is required")

            logger.info("Initializing Gemini image generation service")

            # Configure Gemini API
            genai.configure(api_key=self.config['api_key'])

            # Initialize the model for image generation
            # Note: Using Gemini 2.0 for image generation as specified
            model_name = "gemini-2.0-flash-exp"  # Gemini 2.0 model for image generation
            self.model_instance = genai.GenerativeModel(model_name)

            # Test the connection
            await self._test_connection()

            self.is_initialized = True
            logger.info("Gemini image generation service initialized successfully")

        except Exception as e:
            logger.error(f"Failed to initialize Gemini service: {e}")
            raise RuntimeError(f"Gemini initialization failed: {e}")

    async def _test_connection(self) -> None:
        """Test Gemini API connection."""
        try:
            # Test with a simple text generation to verify connection
            response = await asyncio.get_event_loop().run_in_executor(
                None,
                lambda: self.model_instance.generate_content("Hello")
            )
            logger.info("Gemini API connection test successful")
        except Exception as e:
            logger.error(f"Gemini API connection test failed: {e}")
            raise

    async def generate_image(self, prompt: str, **kwargs) -> Dict[str, Any]:
        """
        Generate image using Gemini AI.

        Args:
            prompt: Text description for image generation
            **kwargs: Additional parameters (size, style, etc.)

        Returns:
            Dictionary containing generated image information
        """
        if not self.is_initialized:
            await self.initialize()

        try:
            logger.info(f"Generating image with prompt: {prompt[:100]}...")

            # Enhance the prompt for better image generation
            enhanced_prompt = self._enhance_prompt(prompt)

            # Generate image using Gemini
            # Note: This is a placeholder implementation as Gemini's image generation
            # API might have specific requirements. Adjust based on actual API.
            response = await self._generate_with_gemini(enhanced_prompt, **kwargs)

            # Process the response
            result = await self._process_image_response(response, prompt)

            logger.info("Image generated successfully")
            return result

        except Exception as e:
            logger.error(f"Image generation failed: {e}")
            return {
                "success": False,
                "error": str(e),
                "image_url": None,
                "image_data": None,
                "metadata": {
                    "prompt": prompt,
                    "error": str(e)
                }
            }

    async def _generate_with_gemini(self, prompt: str, **kwargs) -> Any:
        """
        Generate image using Gemini API.

        Args:
            prompt: Enhanced prompt for image generation
            **kwargs: Additional generation parameters

        Returns:
            Gemini API response
        """
        try:
            # Create image generation prompt
            # Note: Adjust this based on Gemini 2.0's actual image generation API
            image_prompt = f"""Generate a high-quality image based on this description:

            {prompt}

            Please create a detailed, visually appealing image that accurately represents the description.
            Style: {kwargs.get('style', 'realistic')}
            Quality: {kwargs.get('quality', 'high')}
            """

            # Generate content (this might need adjustment based on actual Gemini 2.0 API)
            response = await asyncio.get_event_loop().run_in_executor(
                None,
                lambda: self.model_instance.generate_content(
                    image_prompt,
                    # Add any specific image generation parameters here
                )
            )

            return response

        except Exception as e:
            logger.error(f"Gemini image generation API call failed: {e}")
            raise

    async def _process_image_response(self, response: Any, original_prompt: str) -> Dict[str, Any]:
        """
        Process Gemini API response and extract image data.

        Args:
            response: Gemini API response
            original_prompt: Original user prompt

        Returns:
            Processed image result
        """
        try:
            # Note: This is a placeholder implementation
            # The actual implementation depends on Gemini 2.0's response format

            # For now, we'll create a placeholder response
            # In real implementation, extract image data from response

            if hasattr(response, 'parts') and response.parts:
                # Check if response contains image data
                for part in response.parts:
                    if hasattr(part, 'inline_data'):
                        # Extract image data
                        image_data = part.inline_data.data
                        mime_type = part.inline_data.mime_type

                        return {
                            "success": True,
                            "image_url": None,  # Gemini might not provide URL
                            "image_data": image_data,
                            "mime_type": mime_type,
                            "metadata": {
                                "prompt": original_prompt,
                                "model": self.config.get('model'),
                                "generated_at": asyncio.get_event_loop().time()
                            }
                        }

            # If no image data found, return error
            return {
                "success": False,
                "error": "No image data found in response",
                "image_url": None,
                "image_data": None,
                "metadata": {
                    "prompt": original_prompt,
                    "response_text": str(response.text) if hasattr(response, 'text') else None
                }
            }

        except Exception as e:
            logger.error(f"Failed to process Gemini image response: {e}")
            raise

    def _enhance_prompt(self, prompt: str) -> str:
        """
        Enhance the user prompt for better image generation.

        Args:
            prompt: Original user prompt

        Returns:
            Enhanced prompt
        """
        # Add quality and style enhancements
        enhancements = [
            "high quality",
            "detailed",
            "professional",
            "well-composed"
        ]

        # Check if prompt already contains quality descriptors
        prompt_lower = prompt.lower()
        needed_enhancements = [
            enhancement for enhancement in enhancements
            if enhancement not in prompt_lower
        ]

        if needed_enhancements:
            enhanced = f"{prompt}, {', '.join(needed_enhancements)}"
        else:
            enhanced = prompt

        return enhanced

    def is_available(self) -> bool:
        """
        Check if Gemini service is available.

        Returns:
            True if service is available, False otherwise
        """
        try:
            return bool(self.config.get('api_key')) and self.is_initialized
        except Exception as e:
            logger.error(f"Gemini availability check failed: {e}")
            return False

    async def get_supported_styles(self) -> list[str]:
        """
        Get list of supported image styles.

        Returns:
            List of supported styles
        """
        return [
            "realistic",
            "artistic",
            "cartoon",
            "photographic",
            "digital_art",
            "oil_painting",
            "watercolor",
            "sketch"
        ]

    async def get_model_info(self) -> dict:
        """
        Get information about the Gemini model.

        Returns:
            Dictionary with model information
        """
        return {
            "service_name": self.model_name,
            "model": self.config.get('model', 'gemini-2.0-flash-exp'),
            "capabilities": ["text_to_image", "high_quality_generation"],
            "supported_formats": ["PNG", "JPEG"],
            "initialized": self.is_initialized,
            "available": self.is_available()
        }
