requirements.txt
# Core Framework
streamlit==1.28.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# AI Services
openai==1.3.5
google-generativeai==0.3.2
whisper==1.1.10
transformers==4.35.2
torch==2.1.0
torchaudio==2.1.0

# Web Search
google-search-results==2.4.2

# Audio Processing
soundfile==0.12.1
librosa==0.10.1
pyaudio==0.2.11

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
Pillow==10.1.0

# Logging and Monitoring
loguru==0.7.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
ruff==0.1.6
